+++ var/httpd/htdocs/js/Core.UI.InputFields.js
@@ -1623,7 +1623,16 @@ Core.UI.InputFields = (function (TargetNS) {
                     .on('focus.jstree', '.jstree-anchor', function () {
                         if (!SkipFocus) {
                             Focused = this;
+
+                            // In modernize field selection disable 'backspace' key functionality.
+                            // See bug#14011 (https://bugs.otrs.org/show_bug.cgi?id=14011).
+                            $('.jstree .jstree-anchor').on('keydown', function (e) {
+                                if (e.which === 8 && !$(e.target).is('input')) {
+                                    return false;
+                                }
+                            });
                         } else {
+
                             SkipFocus = false;
                         }
                     })